- Dirty hack to pass in FormElementManager state
