+++ src/ZfcUser/Factory/Form/ChangePassword.php
@@ -6,8 +6,9 @@
  * Time: 9:34 AM
  */
 
-namespace ZfcUser\FormElementManagerFactory\Form;
+namespace ZfcUser\Factory\Form;
 
+use Zend\Form\FormElementManager;
 use Zend\ServiceManager\FactoryInterface;
 use Zend\ServiceManager\ServiceLocatorInterface;
 use ZfcUser\Form;
@@ -16,14 +17,27 @@ class ChangePassword implements FactoryInterface
 {
     public function createService(ServiceLocatorInterface $formElementManager)
     {
-        $sm = $formElementManager->getServiceLocator();
+        if ($formElementManager instanceof FormElementManager) {
+            $sm = $formElementManager->getServiceLocator();
+            $fem = $formElementManager;
+        } else {
+            $sm = $formElementManager;
+            $fem = $sm->get('FormElementManager');
+        }
+
         $options = $sm->get('zfcuser_module_options');
+        $options = clone($options);
+        $options->fem = true;
+
         $form = new Form\ChangePassword(null, $options);
         // Inject the FormElementManager to support custom FormElements
-        $formElementManager = $sm->get('FormElementManager');
-        $form->getFormFactory()->setFormElementManager($formElementManager);
+        $form->getFormFactory()->setFormElementManager($fem);
 
         $form->setInputFilter(new Form\ChangePasswordFilter($options));
+
+        if (!$formElementManager instanceof FormElementManager) {
+            $form->init();
+        }
         return $form;
     }
 }
