+++ src/ZfcUser/Form/ChangeEmail.php
@@ -11,6 +11,10 @@ class ChangeEmail extends ProvidesEventsForm
     {
         $this->setAuthenticationOptions($options);
         parent::__construct($name);
+
+        if (!property_exists($options, "fem")) {
+            $this->init();
+        }
     }
 
     public function init()
