+++ src/ZfcUser/Form/Login.php
@@ -17,6 +17,10 @@ class Login extends ProvidesEventsForm
     {
         $this->setAuthenticationOptions($options);
         parent::__construct($name);
+
+        if (!property_exists($options, "fem")) {
+            $this->init();
+        }
     }
 
     public function init()
