+++ tests/ZfcUserTest/Factory/Form/ChangeEmailFormFactoryTest.php
@@ -3,26 +3,24 @@ namespace ZfcUserTest\Factory\Form;
 
 use Zend\Form\FormElementManager;
 use Zend\ServiceManager\ServiceManager;
-use Zend\Stdlib\Hydrator\ClassMethods;
-use ZfcUser\FormElementManagerFactory\Form\Register as RegisterFactory;
+use ZfcUser\Factory\Form\ChangeEmail as ChangeEmailFactory;
 use ZfcUser\Options\ModuleOptions;
 use ZfcUser\Mapper\User as UserMapper;
 
-class RegisterFormFactoryTest extends \PHPUnit_Framework_TestCase
+class ChangeEmailFormFactoryTest extends \PHPUnit_Framework_TestCase
 {
     public function testFactory()
     {
         $serviceManager = new ServiceManager;
         $serviceManager->setService('zfcuser_module_options', new ModuleOptions);
         $serviceManager->setService('zfcuser_user_mapper', new UserMapper);
-        $serviceManager->setService('zfcuser_register_form_hydrator', new ClassMethods());
 
         $formElementManager = new FormElementManager();
         $formElementManager->setServiceLocator($serviceManager);
         $serviceManager->setService('FormElementManager', $formElementManager);
 
-        $factory = new RegisterFactory();
+        $factory = new ChangeEmailFactory();
 
-        $this->assertInstanceOf('ZfcUser\Form\Register', $factory->createService($formElementManager));
+        $this->assertInstanceOf('ZfcUser\Form\ChangeEmail', $factory->createService($formElementManager));
     }
 }
