+++ tests/ZfcUserTest/Factory/Form/LoginFormFactoryTest.php
@@ -3,24 +3,22 @@ namespace ZfcUserTest\Factory\Form;
 
 use Zend\Form\FormElementManager;
 use Zend\ServiceManager\ServiceManager;
-use ZfcUser\FormElementManagerFactory\Form\ChangeEmail as ChangeEmailFactory;
+use ZfcUser\Factory\Form\Login as LoginFactory;
 use ZfcUser\Options\ModuleOptions;
-use ZfcUser\Mapper\User as UserMapper;
 
-class ChangeEmailFormFactoryTest extends \PHPUnit_Framework_TestCase
+class LoginFormFactoryTest extends \PHPUnit_Framework_TestCase
 {
     public function testFactory()
     {
         $serviceManager = new ServiceManager;
         $serviceManager->setService('zfcuser_module_options', new ModuleOptions);
-        $serviceManager->setService('zfcuser_user_mapper', new UserMapper);
 
         $formElementManager = new FormElementManager();
         $formElementManager->setServiceLocator($serviceManager);
         $serviceManager->setService('FormElementManager', $formElementManager);
 
-        $factory = new ChangeEmailFactory();
+        $factory = new LoginFactory();
 
-        $this->assertInstanceOf('ZfcUser\Form\ChangeEmail', $factory->createService($formElementManager));
+        $this->assertInstanceOf('ZfcUser\Form\Login', $factory->createService($formElementManager));
     }
 }
